# api/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "面向个人房东的租房管理后端系统，支持房源管理、租客管理、水电气收费、租金账单等功能。",
        "title": "租房管理系统 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/properties": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取房东的房源列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房源管理"
                ],
                "summary": "获取房源列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建新房源",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房源管理"
                ],
                "summary": "创建房源",
                "parameters": [
                    {
                        "description": "房源信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_service.CreatePropertyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/properties/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取房源详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房源管理"
                ],
                "summary": "获取房源详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "房源ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新房源信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房源管理"
                ],
                "summary": "更新房源",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "房源ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "房源信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_service.UpdatePropertyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除房源",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房源管理"
                ],
                "summary": "删除房源",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "房源ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/rooms": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取房间列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房间管理"
                ],
                "summary": "获取房间列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "房源ID",
                        "name": "property_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "状态",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "在房源下创建房间",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房间管理"
                ],
                "summary": "创建房间",
                "parameters": [
                    {
                        "description": "房间信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_service.CreateRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/rooms/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "根据ID获取房间详情",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房间管理"
                ],
                "summary": "获取房间详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "房间ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新房间信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房间管理"
                ],
                "summary": "更新房间",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "房间ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "房间信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_service.UpdateRoomRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除房间",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房间管理"
                ],
                "summary": "删除房间",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "房间ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
        "/rooms/{id}/status": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新房间状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "房间管理"
                ],
                "summary": "更新房间状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "房间ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "integer",
                                "format": "int32"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rental-management_internal_pkg_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "rental-management_internal_pkg_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "rental-management_internal_service.CreatePropertyRequest": {
            "type": "object",
            "required": [
                "name",
                "property_type"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "property_type": {
                    "type": "integer"
                }
            }
        },
        "rental-management_internal_service.CreateRoomRequest": {
            "type": "object",
            "required": [
                "property_id",
                "room_number"
            ],
            "properties": {
                "area": {
                    "type": "number"
                },
                "facilities": {
                    "type": "string"
                },
                "floor": {
                    "type": "integer"
                },
                "monthly_rent": {
                    "type": "number"
                },
                "property_id": {
                    "type": "integer"
                },
                "remark": {
                    "type": "string"
                },
                "room_number": {
                    "type": "string"
                }
            }
        },
...
        "rental-management_internal_service.UpdatePropertyRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "rental-management_internal_service.UpdateRoomRequest": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "number"
                },
                "facilities": {
                    "type": "string"
                },
                "floor": {
                    "type": "integer"
                },
                "monthly_rent": {
                    "type": "number"
                },
                "remark": {
                    "type": "string"
                },
                "room_number": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}